requirements.txt
loguru
tqdm
shapely
numpy
pillow
onnxruntime # onnxruntime-gpu
# opencv-contrib-python
opencv-python-headless~=4.10.0.84